requirements.txt
# Core Django
Django==5.2.1
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2024.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Database
dj-database-url==2.1.0
psycopg2-binary==2.9.10

# Environment & Configuration
python-dotenv==1.0.1
python-decouple==3.8

# Image Processing
Pillow==10.0.0

# External Services
requests==2.31.0

# Utilities
python-slugify==8.0.1

# Background Tasks & Caching (Optional for VPS)
celery==5.3.0
redis==4.5.5

# Development & Testing (optional)
django-debug-toolbar==4.2.0
pytest-django==4.7.0
coverage==7.4.0
